+++ Assets/Scripts/Modifiers/Modifier.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 [System.Serializable]
 public class Modifier
 {
-    public int AddedValue;
+    public float AddedValue;
 }
